+++ arch/x86/include/asm/smp.h
@@ -101,12 +101,13 @@ static inline void arch_send_call_function_ipi_mask(const struct cpumask *mask)
 	smp_ops.send_call_func_ipi(mask);
 }
 
-static inline int arch_cpu_physical_id(int cpu)
+/* Physical CPU IDs (APIC IDs) are u64 to match the multikernel core */
+static inline u64 arch_cpu_physical_id(int cpu)
 {
 	return smp_ops.cpu_physical_id(cpu);
 }
 
-static inline int arch_cpu_from_physical_id(int phys_id)
+static inline int arch_cpu_from_physical_id(u64 phys_id)
 {
 	int cpu;
